requirements.txt
absl-py==0.7.1
argh==0.26.2
attrs==19.1.0
backcall==0.1.0
backports.tempfile==1.0
backports.weakref==1.0.post1
bleach==3.1.0
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
Click==7.0
click-plugins==1.1.1
cligj==0.5.0
colorama==0.4.1
ConfigArgParse==0.14.0
cycler==0.10.0
decorator==4.4.0
defusedxml==0.6.0
descartes==1.1.0
docker-pycreds==0.4.0
docutils==0.15.2
easydict==1.9
entrypoints==0.3
experimentlogger-jathr==0.0.2
Fiona==1.8.4
future==0.17.1
GDAL==2.3.3
geopandas==0.4.1
gitdb2==2.0.5
GitPython==2.1.11
gql==0.1.0
graphql-core==2.2
grpcio==1.21.1
h5py==2.9.0
idna==2.8
imageio==2.5.0
ipykernel==5.1.1
ipython==7.6.1
ipython-genutils==0.2.0
ipywidgets==7.5.0
jedi==0.14.1
Jinja2==2.10.1
joblib==0.13.2
jsonschema==3.0.1
jupyter==1.0.0
jupyter-client==5.3.1
jupyter-console==6.0.0
jupyter-core==4.5.0
kiwisolver==1.1.0
mapclassify==2.0.1
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.0
matplotlib2tikz==0.7.4
mistune==0.8.4
mkl-fft==1.0.12
mkl-random==1.0.2
munch==2.3.2
nbconvert==5.5.0
nbformat==4.4.0
networkx==2.3
nose==1.3.7
notebook==6.0.0
numpy==1.16.4
nvidia-ml-py3==7.352.0
olefile==0.46
pandas==0.24.2
pandocfilters==1.4.2
parso==0.5.1
pathtools==0.1.2
patsy==0.5.1
pickleshare==0.7.5
Pillow==6.0.0
pkginfo==1.5.0.1
prometheus-client==0.7.1
promise==2.2.1
prompt-toolkit==2.0.9
protobuf==3.8.0
psutil==5.6.3
psycopg2==2.7.6.1
pycparser==2.19
Pygments==2.4.2
pymongo==3.8.0
pyparsing==2.4.0
pyproj==1.9.6
pyrsistent==0.15.3
python-dateutil==2.8.0
pytz==2019.1
PyWavelets==1.0.3
pywinpty==0.5.5
PyYAML==5.1.1
pyzmq==18.0.2
qtconsole==4.5.2
readme-renderer==24.0
requests==2.22.0
requests-toolbelt==0.9.1
Rtree==0.8.3
Rx==1.6.1
scikit-image==0.15.0
scikit-learn==0.21.2
scipy==1.3.0
seaborn==0.9.0
Send2Trash==1.5.0
sentry-sdk==0.10.1
Shapely==1.6.4.post1
shortuuid==0.5.0
simplejson==3.16.0
six==1.12.0
smmap2==2.0.5
SQLAlchemy==1.3.5
statsmodels==0.10.1
subprocess32==3.5.4
tensorboard==1.14.0
terminado==0.8.2
testpath==0.4.2
torch==1.1.0
torchsummary==1.5.1
torchvision==0.3.0
tornado==6.0.3
tqdm==4.32.2
traitlets==4.3.2
twine==1.13.0
urllib3==1.25.3
wandb==0.8.5
watchdog==0.9.0
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.15.4
widgetsnbextension==3.5.0
wincertstore==0.2
